complete review fixes: move header to include/rpc, fix includes + resolver call site
